% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_base_eu_2025.R
\docType{data}
\name{base_eu_2025}
\alias{base_eu_2025}
\title{base_eu_2025}
\format{
A data frame with 295 rows and 11 variables:
\describe{
  \item{ pays }{  character }
  \item{ region }{  character }
  \item{ annee }{  integer }
  \item{ Population }{  integer }
  \item{ PIB }{  numeric }
  \item{ PIB_habitant }{  numeric }
  \item{ Taux_Croissance }{  numeric }
  \item{ Taux_Chomage }{  numeric }
  \item{ Inflation }{  numeric }
  \item{ Dette_Publique }{  numeric }
  \item{ Balance_Commerciale }{  numeric }
}
}
\source{
Source
}
\usage{
base_eu_2025
}
\description{
Description.
}
\keyword{datasets}
