% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2_find.R
\name{chi2_find}
\alias{chi2_find}
\title{Find modalities related to a criterion}
\usage{
chi2_find(df, criterion)
}
\arguments{
\item{df}{data.frame}

\item{criterion}{character string: criterion that spots target rows}
}
\value{
data.frame
}
\description{
Find modalities related to a criterion
}
