% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coltypes}
\alias{coltypes}
\title{coltypes()}
\usage{
coltypes(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
vector
}
\description{
Create vector of df's column types. Similar to colnames(), but with column types instead of names.
}
