% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_recursion_depth}
\alias{get_recursion_depth}
\title{get_recursion_depth}
\usage{
get_recursion_depth(x, depth = 0)
}
\arguments{
\item{x}{: input list}

\item{depth}{: depth of x in another list (1 if x in a list. 2 if x is in a list of lists. Etc.)}
}
\value{
integer
}
\description{
get recursion depth of a list
}
