% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tac.R
\name{get_tac_column}
\alias{get_tac_column}
\title{Contingency table for column `col_name` in data.frame `df}
\usage{
get_tac_column(df, col_name, values, strates)
}
\arguments{
\item{df}{Input data.frame}

\item{col_name}{string : name of column to which generate the contingency table}

\item{values}{Vector of columns that serve as measures (amounts, counts, etc.)}

\item{strates}{Vector of column names by which to stratify the contingency tables}
}
\description{
Contingency table for column `col_name` in data.frame `df
}
