% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.Date}
\alias{is.Date}
\title{is.Date}
\usage{
is.Date(x)
}
\arguments{
\item{x}{object}
}
\value{
TRUE/FALSE
}
\description{
Returns TRUE or FALSE depending on whether its argument is of Date type or not
}
