% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.POSIXct}
\alias{is.POSIXct}
\title{is.POSIXct}
\usage{
is.POSIXct(x)
}
\arguments{
\item{x}{object}
}
\value{
TRUE/FALSE
}
\description{
Returns TRUE or FALSE depending on whether its argument is of POSIXct type or not
}
