% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.POSIXlt}
\alias{is.POSIXlt}
\title{is.POSIXlt}
\usage{
is.POSIXlt(x)
}
\arguments{
\item{x}{object}
}
\value{
TRUE/FALSE
}
\description{
Returns TRUE or FALSE depending on whether its argument is of POSIXlt type or not
}
