% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.POSIXt}
\alias{is.POSIXt}
\title{is.POSIXt}
\usage{
is.POSIXt(x)
}
\arguments{
\item{x}{object}
}
\value{
TRUE/FALSE
}
\description{
Returns TRUE or FALSE depending on whether its argument is of POSIXxt type or not
}
