% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sums.R
\name{sums}
\alias{sums}
\title{sums: Generate a synthetic table of sums of numeric columns}
\usage{
sums(tab)
}
\arguments{
\item{tab}{data.frame: Input data.frame}
}
\value{
data.frame
}
\description{
Get a synthetic table of sums of all numeric columns of a data.frame
}
\examples{
sums(mtcars)  # Checking NA values for all columns of mtcars (none)

}
