% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tac.R
\name{tac}
\alias{tac}
\title{Computes a contingency table (tac) of all columns in a dataframe for control purposes}
\usage{
tac(
  df,
  values = NULL,
  sample_rate = 0.01,
  force_identifier = "NULL",
  force_levels = "NULL",
  force_quanti = "NULL",
  strates = NULL
)
}
\arguments{
\item{df}{Input data.frame}

\item{values}{Vector of columns that serve as measures (amounts, counts, etc.)}

\item{sample_rate}{Sampling rate, if df is a remote table}

\item{force_identifier}{Vector of names of columns that user wants to be considred as identifiers}

\item{force_levels}{Vector of names of numeric columns with discrete modalities (not continuous)}

\item{force_quanti}{Vector of names of numeric columns that are continuous (not discrete)}

\item{strates}{Vector of column names by which to stratify the contingency tables}
}
\value{
data.frame
}
\description{
Contingency table (tac) of all columns in a dataframe for control purposes
}
\examples{
tab <- tac(iris) # calculate column frequencies

}
