% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{TAC-based Outlier Control (TOC)}
\usage{
toc(
  df1,
  df2,
  values = NULL,
  a = 10,
  r = 0.34,
  sample_rate = 0.01,
  force_identifier = "NULL",
  force_levels = "NULL",
  force_quanti = "NULL"
)
}
\arguments{
\item{df1}{Input data.frame (to compare with df2)}

\item{df2}{Input data.frame (to compare with df1)}

\item{values}{Vector of columns that serve as measures (amounts, counts, etc.)}

\item{a}{Allowed absolute variation}

\item{r}{Allowed relative variation}

\item{sample_rate}{Sampling rate, if df is a remote table}

\item{force_identifier}{Vector of names of columns that user wants to be considred as identifiers}

\item{force_levels}{Vector of names of numeric columns with discrete modalities (not continuous)}

\item{force_quanti}{Vector of names of numeric columns that are continuous (not discrete)
#'}
}
\value{
data.frame
}
\description{
Generalized detection of outlier values in a database, based on contingency tables (tac)
}
\examples{
verif <- toc(base_eu_2024, base_eu_2025[base_eu_2025$annee != 2025, ]) 
# ecart  significatif sur la modalite "Catalogne" (disparition dans la base 2024)

}
