% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc_score}
\alias{toc_score}
\title{Scoring significativity of difference between two values x and y}
\usage{
toc_score(x, y, a)
}
\arguments{
\item{x}{(num) First value to compare}

\item{y}{(num) Second value to compare}

\item{a}{(num) Absolute difference threshold below which all differences are considered normal}
}
\value{
numeric
}
\description{
Difference score between x and y (0 = no significant difference, >0 = presence of significant difference)
}
\examples{
toc_score(15, 1500, a = 500) # 1.91: significant difference
toc_score(1432, 1501, a = 100) # 0: non-significant difference

}
