% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.R
\name{miss}
\alias{miss}
\title{Missing: Generate a synthetic table of missing values for all columns of a data.frame}
\usage{
miss(df, values = NULL, view = FALSE)
}
\arguments{
\item{df}{data.frame: Input data.frame}

\item{values}{column: Variable (~weight) to measure the number of missing values (otherwise, count of rows)}

\item{view}{boolean: Display a glimpse of cases with NA values}
}
\value{
data.frame
}
\description{
Get a synthetic table of missing values for all columns of a data.frame
}
\examples{
miss(mtcars)  # Checking NA values for all columns of mtcars (none)

}
