% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{TAC-based Outlier Control (TOC)}
\usage{
toc(
  df1,
  df2,
  values = NULL,
  a = 10,
  r = 0.34,
  sample_rate = 0.01,
  num_but_discrete = "NULL"
)
}
\arguments{
\item{df1}{Input data.frame (to compare with df2)}

\item{df2}{Input data.frame (to compare with df1)}

\item{values}{Vector of columns that serve as measures (amounts, counts, etc.)}

\item{a}{Allowed absolute variation}

\item{r}{Allowed relative variation}

\item{sample_rate}{Sampling rate, if df is a remote table}

\item{num_but_discrete}{Numeric variables to be treated as discrete modal variables. If 'all', all numeric variables are treated as discrete modal variables.}
}
\value{
data.frame
}
\description{
Generalized detection of outlier values in a database, based on contingency tables (tac)
}
